return http.Server when TLS config is missing
The common pattern in the implementation of Server (including the
http.Server and https.Server objects from the Node.js core) is
the following (note that `this` is completely ignored):

    if (!(this instanceof Server)) return new Server(...arguments);

There are three relevant `Server` implementations, with the following
inheritance tree:

    - tls.Server
     +-- http.Server
     +-- https.Server
       +-- httpolyglot.Server

When `httpolyglot.Server` is constructed without TLS config, it falls
back to constructing itself using `http.Server`:

    // this is an instance of httpolyglot.Server
    http.Server.call(this, requestListener);

But `this` is not an instance of `http.Server`, so a new `http.Server`
instance is created (and discarded) and `this` is not modified.
So, when `listen()` is called, the server starts listening on a port
but does not have any request handlers. Consequently, when a http(s)
request is sent to the server, the connection is accepted but the
request is never processed.

To fix this, be explicit and return the new instance of `http.Server`.